feat(qual): add reference production lab
Make readiness reproducible across HA, PITR, SDK, perimeter, HTTP/2, replay, and soak qualification.

+++ internal/statebolt/maintenance.go
@@ -19,7 +19,7 @@ import (
 // refuse a backup whose hash no longer matches its sidecar.
 type RecoveryManifest struct {
 	FormatVersion             int       `json:"format_version"`
-	SchemaVersion             int       `json:"schema_version"`
+	SchemaVersion             uint64    `json:"schema_version"`
 	DatabaseSHA256            string    `json:"database_sha256"`
 	DatabaseBytes             int64     `json:"database_bytes"`
 	PolicyID                  string    `json:"policy_id,omitempty"`
@@ -114,7 +114,7 @@ func (s *Store) BackupWithRecoveryManifest(path, manifestPath string, metadata R
 	if manifestPath == "" {
 		manifestPath = path + ".manifest.json"
 	}
-	data, err := os.ReadFile(path)
+	data, err := os.ReadFile(path) // #nosec G304 -- backup path is an operator-owned maintenance input.
 	if err != nil {
 		return err
 	}
@@ -143,11 +143,11 @@ func ValidateRecoveryManifest(backup, manifestPath string) error {
 	if backup == "" || manifestPath == "" {
 		return errors.New("statebolt: backup and recovery manifest paths required")
 	}
-	data, err := os.ReadFile(backup)
+	data, err := os.ReadFile(backup) // #nosec G304 -- backup path is validated by the recovery workflow.
 	if err != nil {
 		return err
 	}
-	manifestData, err := os.ReadFile(manifestPath)
+	manifestData, err := os.ReadFile(manifestPath) // #nosec G304 -- manifest path is an operator-owned recovery input.
 	if err != nil {
 		return err
 	}
@@ -275,7 +275,7 @@ func atomicWriteRestricted(path string, data []byte) (err error) {
 		return err
 	}
 	tmpName = ""
-	d, err := os.Open(dir)
+	d, err := os.Open(dir) // #nosec G304 -- directory is derived from an operator-owned recovery path.
 	if err != nil {
 		return err
 	}
@@ -346,7 +346,7 @@ func restoreBackup(backup, target string) (err error) {
 		_ = tmp.Close()
 		return err
 	}
-	src, err := os.Open(backup)
+	src, err := os.Open(backup) // #nosec G304 -- backup path is validated by the recovery workflow.
 	if err != nil {
 		_ = tmp.Close()
 		return err
@@ -372,7 +372,7 @@ func restoreBackup(backup, target string) (err error) {
 		return err
 	}
 	tmpName = ""
-	d, err := os.Open(dir)
+	d, err := os.Open(dir) // #nosec G304 -- directory is derived from an operator-owned recovery path.
 	if err != nil {
 		return err
 	}
@@ -433,7 +433,7 @@ func CompactFile(path string) (err error) {
 		return err
 	}
 	tmpName = ""
-	d, err := os.Open(dir)
+	d, err := os.Open(dir) // #nosec G304 -- directory is derived from an operator-owned recovery path.
 	if err != nil {
 		return err
 	}
