+++ internal/statebolt/operator.go
@@ -108,12 +108,12 @@ func appendOperatorTx(tx *bolt.Tx, rec control.OperatorRecord) error {
 	audit := tx.Bucket(bucketOperatorAudit)
 	seq := btoi(audit.Get(keyAuditSequence))
 	seq++
-	rec.Sequence = uint64(seq)
+	rec.Sequence = seq
 	env, err := json.Marshal(persistedOperatorRecord{SchemaVersion: operatorRecordSchemaVersion, Record: rec})
 	if err != nil {
 		return err
 	}
-	if err := audit.Put(itob(uint64(seq)), env); err != nil {
+	if err := audit.Put(itob(seq), env); err != nil {
 		return err
 	}
 	return audit.Put(keyAuditSequence, itob(uint64(seq)))
@@ -284,14 +284,14 @@ func (s *Store) ListOperatorAudit(after uint64, limit int) ([]control.OperatorRe
 				return fmt.Errorf("statebolt: corrupt operator audit key: %w", ErrMigrationRequired)
 			}
 			seq := btoi(k)
-			if uint64(seq) <= after {
+			if seq <= after {
 				continue
 			}
 			var p persistedOperatorRecord
 			if err := json.Unmarshal(v, &p); err != nil || p.SchemaVersion != operatorRecordSchemaVersion {
 				return fmt.Errorf("statebolt: corrupt operator audit record: %w", ErrMigrationRequired)
 			}
-			p.Record.Sequence = uint64(seq)
+			p.Record.Sequence = seq
 			out = append(out, p.Record)
 		}
 		return nil
